pyproject.toml
[tool.poetry]
name = "todo-sub"
version = "0.1.0"
description = ""
authors = ["aasifshahzad <asif17186@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
sqlmodel = "^0.0.16"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
